requirements.txt
PyYAML==6.0.1
rospkg==1.5.0
nicegui==1.4.20
pillow==10.3.0
tornado==6.4
simplejpeg==1.7.2